Refactor directory creation and add error handling for PermissionError

+++ main.py
@@ -1,22 +1,28 @@
 import os
 import shutil
+import time
 
 from matplotlib import pyplot as plt
 
 from learn import Learn
+from settings import *
 
 os.environ["KMP_DUPLICATE_LIB_OK"] = "TRUE"
-
-rule_number = 30
-path = f"results/rule_{rule_number}/"
 
 
 def make_dir(path):
     try:
-        shutil.rmtree(path)
-    except:
-        pass
-    os.makedirs(path)
+        if os.path.exists(path):
+            shutil.rmtree(path)
+        else:
+            os.makedirs(path)
+    except PermissionError as e:
+        print(f"PermissionError encountered: {e}. Retrying...")
+        time.sleep(1)  # Wait a bit in case the file is temporarily locked
+        if os.path.exists(path):  # Check again to see if the directory still exists
+            shutil.rmtree(path)
+        else:
+            os.makedirs(path)
 
 
 def plot_automata(rule_number, automata, path):
@@ -28,20 +34,25 @@ def plot_automata(rule_number, automata, path):
     plt.close()
 
 
-for i in range(141, 256):
-    rule_number = i
-    path = f"results/rule_{rule_number}/"
-    make_dir(path)
-    learn = Learn(rule_number)
-    real_automata = learn.automata.generate(100)
-    plot_automata(rule_number, real_automata, path)
-    learn.train(path=path)
+# for i in range(141, 256):
+#     rule_number = i
+#     path = f"results/rule_{rule_number}/"
+#     make_dir(path)
+#     learn = Learn(rule_number)
+#     real_automata = learn.automata.generate(100)
+#     plot_automata(rule_number, real_automata, path)
+#     learn.train(path=path)
 
-
-# rule_number = 30
-# path = f"results/rule_{rule_number}/"
-# make_dir(path)
-# learn = Learn(rule_number)
+make_dir(PATH)
+learn = Learn(
+    rule_number=RULE_NUMBER,
+    num_cells=NUM_CELLS,
+    num_generations=GENERATIONS,
+    learning_rate=LEARNING_RATE,
+    training_size=TRAINING_SIZE,
+    epochs=EPOCHS,
+    path=PATH,
+)
 # real_automata = learn.automata.generate(100)
 # plot_automata(rule_number, real_automata, path)
-# learn.train(path=path)
+learn.train()
